ERC20ConfidentialUpgradeable version with tests
