+++ .changeset/rare-cobras-run.md
@@ -1,0 +1,5 @@
+---
+"fhenix-confidential-contracts": patch
+---
+
+ERC20ConfidentialUpgradeable version with tests
